move testcases to correct file
